Module Signatures

This module contains all module signatures that the user could use to build an ocamlbuild plugin.

module type OrderedTypePrintable = sig ... end

module type SET = sig ... end

module type LIST = sig ... end

module type STRING = sig ... end

module type TAGS = sig ... end

module type PATHNAME = sig ... end

module type COMMAND = sig ... end

Provides an abstract type for easily building complex shell commands without making quotation mistakes.

module type GLOB = sig ... end

A self-contained module implementing extended shell glob patterns who have an expressive power equal to boolean combinations of regular expressions.

module type LOG = sig ... end

Module for modulating the logging output with the logging level.

module type OUTCOME = sig ... end

module type MISC = sig ... end

module type OPTIONS = sig ... end

module type ARCH = sig ... end

module type FINDLIB = sig ... end

Findlib / Ocamlfind tools.

module type PLUGIN = sig ... end

This module contains the functions and values that can be used by plugins.